Fix USBController destructor hang with m_is_disconnected flag
Now that the destructor crash is fixed, it hangs instead. The crash
has probably always masked this issue. I don't understand libusb much
as it isn't well documented but it seems that by the time we check
that all transfers have been cancelled, new ones have already been
submitted. Setting m_is_disconnected to true beforehand to indicate we
are destructing appears to remedy this.